feat: add proper permissions for collection/collectionsong views

+++ src/backend/musikk/streaming/api/v1/views/collections.py
@@ -1,11 +1,9 @@
 from django.db import transaction
-from rest_framework import status, serializers
+from rest_framework import status
 from rest_framework.generics import ListAPIView, RetrieveAPIView, get_object_or_404
 from rest_framework.parsers import MultiPartParser, FormParser
-from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from rest_framework.exceptions import NotFound
 
 from websockets.event_helpers import send_ws_event
 from streaming.api.v1.serializers.collections import (
@@ -13,12 +11,12 @@ from streaming.api.v1.serializers.collections import (
     CollectionSerializerDetailed,
     CollectionCreateSerializer,
 )
-from streaming.models.collections import Collection
+from streaming.models.collections import Collection, CollectionType
 from streaming.models.songs import CollectionSong
+from streaming.permissions import IsPublicOrCollectionAuthor
 
 
-class CollectionLatestView(ListAPIView):
-    permission_classes = [IsAuthenticated]
+class PlaylistsLatestView(ListAPIView):
     queryset = Collection.objects.all()
     serializer_class = CollectionSerializerBasic
     amount = 50
@@ -27,6 +25,24 @@ class CollectionLatestView(ListAPIView):
         qs = (
             super()
             .get_queryset()
+            .filter(type=CollectionType.PLAYLIST)
+            .exclude(private=True)
+            .order_by("-date_added")[: self.amount]
+        )
+        return qs
+
+
+class AlbumsLatestView(ListAPIView):
+
+    queryset = Collection.objects.all()
+    serializer_class = CollectionSerializerBasic
+    amount = 50
+
+    def get_queryset(self):
+        qs = (
+            super()
+            .get_queryset()
+            .filter(type=CollectionType.ALBUM)
             .exclude(private=True)
             .order_by("-date_added")[: self.amount]
         )
@@ -34,7 +50,6 @@ class CollectionLatestView(ListAPIView):
 
 
 class CollectionPersonalView(APIView):
-    permission_classes = [IsAuthenticated]
     serializer_class = CollectionSerializerBasic
 
     def get(self, request, *args, **kwargs):
@@ -65,38 +80,37 @@ class CollectionPersonalView(APIView):
 
 class CollectionRetrieveView(RetrieveAPIView):
     lookup_field = "uuid"
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsPublicOrCollectionAuthor]
     queryset = Collection.objects.all()
     serializer_class = CollectionSerializerBasic
 
 
 class CollectionDetailView(RetrieveAPIView):
     lookup_field = "uuid"
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsPublicOrCollectionAuthor]
     queryset = Collection.objects.all()
     serializer_class = CollectionSerializerDetailed
 
 
 class CollectionAddLikedView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsPublicOrCollectionAuthor]
 
     def post(self, request, *args, **kwargs):
-        user = self.request.user
-        user_uuid = user.uuid
-
         collection_uuid = kwargs["uuid"]
         with transaction.atomic():
             collection = get_object_or_404(Collection, uuid=collection_uuid)
-            user.streamingprofile.followed_collections.add(collection)
+            self.check_object_permissions(request, collection)
+            self.request.user.streamingprofile.followed_collections.add(collection)
 
+        # TODO: move?
         send_ws_event(
-            f"user_{user_uuid}",
+            f"user_{self.request.user.uuid}",
             event_handler="base.event",
             event_name="invalidate.query",
             query_key=["openCollection"],
         )
         send_ws_event(
-            f"user_{user_uuid}",
+            f"user_{self.request.user.uuid}",
             event_handler="base.event",
             event_name="invalidate.query",
             query_key=["collectionsPersonal"],
@@ -105,18 +119,15 @@ class CollectionAddLikedView(APIView):
 
 
 class CollectionRemoveSong(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsPublicOrCollectionAuthor]
 
     def delete(self, request, *args, **kwargs):
-        collection_uuid = kwargs["collection_uuid"]
-        collection_song_uuid = kwargs["song_uuid"]
-
         collection_song = get_object_or_404(
             CollectionSong,
-            collection__uuid=collection_uuid,
-            uuid=collection_song_uuid,
-            author=request.user,
+            collection__uuid=kwargs["collection_uuid"],
+            uuid=kwargs["song_uuid"],
         )
+        self.check_object_permissions(request, collection_song)
 
         collection_song.delete()
         send_ws_event(
@@ -131,19 +142,14 @@ class CollectionRemoveSong(APIView):
 
 
 class CollectionAddSong(APIView):
-    permission_classes = [IsAuthenticated]
-
     def post(self, request, *args, **kwargs):
-        collection_uuid = kwargs["collection_uuid"]
-        collection_song_uuid = kwargs["song_uuid"]
-
         user = request.user
         with transaction.atomic():
             collection = get_object_or_404(
-                Collection, uuid=collection_uuid, authors__in=user
+                Collection, uuid=kwargs["collection_uuid"], authors__in=user
             )
             collection_song = get_object_or_404(
-                CollectionSong, uuid=collection_song_uuid
+                CollectionSong, uuid=kwargs["song_uuid"]
             )
 
             CollectionSong.objects.create(
@@ -159,7 +165,6 @@ class CollectionCreateView(APIView):
     """
 
     parser_classes = [MultiPartParser, FormParser]
-    permission_classes = [IsAuthenticated]
 
     def post(self, request, *args, **kwargs):
         data = request.data.copy()
@@ -172,45 +177,38 @@ class CollectionCreateView(APIView):
         if not isinstance(songs, list):
             songs = [songs]
 
-        payload = {
-            "title": data.get("title"),
-            "description": data.get("description", ""),
-            "image": request.FILES.get("image"),
-            "authors": authors,
-            "songs": songs,
-            "type": data.get("type"),
-        }
         serializer = CollectionCreateSerializer(
-            data=payload, context={"request": request}
+            data={
+                "title": data["title"],
+                "description": data.get("description", ""),
+                "image": request.FILES.get("image"),
+                "authors": authors,
+                "songs": songs,
+                "type": data["type"],
+            },
+            context={"request": request},
         )
-        try:
-            serializer.is_valid(raise_exception=True)
-        except serializers.ValidationError as ex:
-            return Response({"failed": ex.detail}, status=status.HTTP_400_BAD_REQUEST)
 
+        serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(status=status.HTTP_201_CREATED)
 
 
 class AlbumBySongView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsPublicOrCollectionAuthor]
 
     def get(self, request, *args, **kwargs):
-        collection_song_uuid = kwargs.get("song_uuid")
+        collection_song = get_object_or_404(
+            CollectionSong.objects.select_related("song"),
+            uuid=kwargs["song_uuid"],
+        )
 
-        with transaction.atomic():
-            collection_song = get_object_or_404(
-                CollectionSong, uuid=collection_song_uuid
-            )
-            album_song = CollectionSong.objects.filter(
-                collection__type="album", song__uuid=collection_song.song.uuid
-            )
-            if not album_song:
-                raise NotFound(f"Album for song {collection_song.song.uuid} not found.")
+        album = get_object_or_404(
+            Collection,
+            type=CollectionType.ALBUM,
+            base_songs=collection_song.song,
+        )
+        self.check_object_permissions(request, album)
 
-            album = get_object_or_404(
-                Collection, type="album", uuid=album_song[0].collection.uuid
-            )
-
-        album = CollectionSerializerBasic(album, context={"request": request}).data
-        return Response(status=status.HTTP_200_OK, data=album)
+        data = CollectionSerializerBasic(album, context={"request": request}).data
+        return Response(status=status.HTTP_200_OK, data=data)
